Dedupe _remaining_ms and wall_seconds budget constants (#1550)
Extract the duplicated remaining_ms computation into a shared
_budget module in graph_optimization. Replace 2 inline duplications
in operations.py and invariants.py with calls to the shared helper.
The 4th copy in _independence_z3.py is left as-is due to the
native-math-boundary constraint preventing jacobian.math from
importing jacobian.domains.
